crawlers/requirements.txt
beautifulsoup4==4.8.2
black==19.10b0
bs4==0.0.1
pymongo==3.9.0
Scrapy==1.7.3
pytest==5.3.5
ipdb==0.12.3
pytest-cov==2.8.1
python-dotenv==0.11.0